chore: add sampling as a transform
