+++ __tests__/integration/charts/unemployment-area-stacked-sampling.ts
@@ -7,20 +7,19 @@ export function unemploymentAreaStackedSampling(): G2Spec {
     data: {
       type: 'fetch',
       value: 'data/unemployment-by-industry.csv',
-      transform: [
-        {
-          type: 'sampling',
-          thresholds: 30,
-          groupBy: 'industry',
-          strategy: 'max',
-          fields: ['unemployed'],
-        },
-      ],
     },
     children: [
       {
         type: 'area',
-        transform: [{ type: 'stackY' }],
+        transform: [
+          { type: 'stackY' },
+          {
+            type: 'sampling',
+            thresholds: 30,
+            groupBy: 'color',
+            y: 'max',
+          },
+        ],
         encode: {
           x: 'date',
           y: 'unemployed',
