+++ __tests__/unit/stdlib/index.spec.ts
@@ -248,7 +248,6 @@ import {
   Arc as ArcTransform,
   WordCloud,
   Join,
-  Sampling,
 } from '../../../src/data';
 import {
   HideOverlap,
@@ -277,7 +276,6 @@ describe('stdlib', () => {
       'data.arc': ArcTransform,
       'data.wordCloud': WordCloud,
       'data.join': Join,
-      'data.sampling': Sampling,
       // 'transform.voronoi': Voronoi,
       'transform.maybeZeroY1': MaybeZeroY1,
       'transform.maybeZeroX': MaybeZeroX,
@@ -319,6 +317,7 @@ describe('stdlib', () => {
       'transform.maybeFunctionAttribute': MaybeFunctionAttribute,
       'transform.maybeGradient': MaybeGradient,
       'transform.pack': Pack,
+      'transform.sampling': Sampling,
       'coordinate.cartesian': Cartesian,
       'coordinate.polar': Polar,
       'coordinate.helix': Helix,
