+++ src/spec/data.ts
@@ -1,4 +1,4 @@
-import { DataComponent, Vector2 } from '../runtime';
+import { DataComponent, Primitive, Vector2 } from '../runtime';
 export type Data = FetchConnector | InlineConnector | any;
 
 export type FetchConnector = {
@@ -27,7 +27,6 @@ export type DataTransform =
   | JoinTransform
   | FilterTransform
   | MapTransform
-  | SamplingTransform
   | CustomTransform;
 
 export type DataTransformTypes =
@@ -39,7 +38,6 @@ export type DataTransformTypes =
   | 'join'
   | 'filter'
   | 'map'
-  | 'sampling'
   | 'custom'
   | DataComponent;
 
@@ -257,22 +255,4 @@ export type TreeTransform = Omit<ClusterTransform, 'type'> & {
 export type MapTransform = {
   type?: 'map';
   callback?: (d: any) => any;
-};
-
-export type SamplingTransform = {
-  type?: 'sampling';
-  thresholds?: number;
-  groupBy?: string | string[];
-  fields: string | string[];
-  strategy?:
-    | 'none'
-    | 'lttb'
-    | 'mean'
-    | 'median'
-    | 'max'
-    | 'min'
-    | 'sum'
-    | 'first'
-    | 'last'
-    | ((frames: number[]) => number);
 };
