+++ src/stdlib/index.ts
@@ -230,6 +230,7 @@ import {
   MaybeFunctionAttribute,
   MaybeGradient,
   MaybeTooltip,
+  Sampling,
 } from '../transform';
 import {
   Fetch,
@@ -248,7 +249,6 @@ import {
   Arc as ArcTransform,
   WordCloud,
   Join,
-  Sampling,
 } from '../data';
 import { HideOverlap, DodgeY, Spider, Surround } from '../label-transform';
 import { geoLibrary } from '../geo';
@@ -271,7 +271,6 @@ export function createLibrary(): G2Library {
     'data.arc': ArcTransform,
     'data.wordCloud': WordCloud,
     'data.join': Join,
-    'data.sampling': Sampling,
     // 'transform.voronoi': Voronoi,
     'transform.maybeZeroY1': MaybeZeroY1,
     'transform.maybeZeroX': MaybeZeroX,
@@ -313,6 +312,7 @@ export function createLibrary(): G2Library {
     'transform.maybeGradient': MaybeGradient,
     'transform.maybeTooltip': MaybeTooltip,
     'transform.pack': Pack,
+    'transform.sampling': Sampling,
     'coordinate.cartesian': Cartesian,
     'coordinate.polar': Polar,
     'coordinate.helix': Helix,
