+++ src/transform/index.ts
@@ -38,6 +38,7 @@ export { FlexX } from './flexX';
 export { Pack } from './pack';
 export { BinX } from './binX';
 export { Bin } from './bin';
+export { Sampling } from './sampling';
 
 export type { MaybeZeroY1Options } from './maybeZeroY1';
 export type { MaybeStackYOptions } from './maybeStackY';
@@ -79,3 +80,4 @@ export type { FlexXOptions } from './flexX';
 export type { PackOptions } from './pack';
 export type { BinXOptions } from './binX';
 export type { BinOptions } from './bin';
+export type { SamplingOptions } from './sampling';
